+++ src/main/java/com/example/springrestapi/repository/CloudVendorRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CloudVendorRepository extends JpaRepository<CloudVendorEntity, Long> {
+public interface CloudVendorRepository extends JpaRepository<CloudVendorEntity, Integer> {
     List<CloudVendorEntity> findByVendorName(String name);
 //    CloudVendorEntity findByVendorName(String name); //Return unique value in db. If value in db > 1 => Error
 }
