+++ src/main/java/com/example/springrestapi/service/ICloudVendorService.java
@@ -6,9 +6,9 @@ import java.util.List;
 
 public interface ICloudVendorService {
     public CloudVendorDTO create(CloudVendorDTO cloudVendorDTO);
-    public CloudVendorDTO update(long id, CloudVendorDTO cloudVendorDTO);
-    public void delete(long id);
-    public CloudVendorDTO getOne(long id);
+    public CloudVendorDTO update(int id, CloudVendorDTO cloudVendorDTO);
+    public void delete(int id);
+    public CloudVendorDTO getOne(int id);
     public List<CloudVendorDTO> getAll();
 
     List<CloudVendorDTO> getByVendorName(String name);
